pyproject.toml
[project]
name = "html2md-skill"
version = "0.1.0"
description = "HTML 页面转 Markdown 的通用 skill，面向论文类材料"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "laotang" }]
dependencies = [
    "httpx>=0.27",
    "lxml>=5.2",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8.1",
    "pydantic>=2.7",
    "python-ulid>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]
browser = [
    "playwright>=1.45",
]

[project.scripts]
html2md-skill = "html2md_skill.__main__:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/html2md_skill"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"
